fix show bug
