+++ mmdet/core/mask/__init__.py
@@ -1,7 +1,8 @@
 from .mask_target import mask_target
 from .structures import BitmapMasks, PolygonMasks
-from .utils import split_combined_polys
+from .utils import encode_mask_results, split_combined_polys
 
 __all__ = [
-    'split_combined_polys', 'mask_target', 'BitmapMasks', 'PolygonMasks'
+    'split_combined_polys', 'mask_target', 'BitmapMasks', 'PolygonMasks',
+    'encode_mask_results'
 ]
