+++ mmdet/models/detectors/cascade_rcnn.py
@@ -31,4 +31,4 @@ class CascadeRCNN(TwoStageDetector):
         else:
             if isinstance(result, dict):
                 result = result['ensemble']
-        super(CascadeRCNN, self).show_result(data, result, **kwargs)
+        return super(CascadeRCNN, self).show_result(data, result, **kwargs)
