+++ tools/test_robustness.py
@@ -16,7 +16,7 @@ from robustness_eval import get_results
 
 from mmdet import datasets
 from mmdet.apis import set_random_seed
-from mmdet.core import eval_map, wrap_fp16_model
+from mmdet.core import encode_mask_results, eval_map, wrap_fp16_model
 from mmdet.datasets import build_dataloader, build_dataset
 from mmdet.models import build_detector
 
@@ -98,10 +98,16 @@ def single_gpu_test(model, data_loader, show=False):
     for i, data in enumerate(data_loader):
         with torch.no_grad():
             result = model(return_loss=False, rescale=not show, **data)
-        results.append(result)
 
         if show:
             model.module.show_result(data, result, dataset.img_norm_cfg)
+
+        # encode mask results
+        if isinstance(result, tuple):
+            bbox_results, mask_results = result
+            encoded_mask_results = encode_mask_results(mask_results)
+            result = bbox_results, encoded_mask_results
+        results.append(result)
 
         batch_size = data['img'][0].size(0)
         for _ in range(batch_size):
@@ -119,6 +125,13 @@ def multi_gpu_test(model, data_loader, tmpdir=None):
     for i, data in enumerate(data_loader):
         with torch.no_grad():
             result = model(return_loss=False, rescale=True, **data)
+            # encode mask results
+            if isinstance(result, tuple):
+                bbox_results, mask_results = result
+                encoded_mask_results = encode_mask_results(mask_results)
+                result = bbox_results, encoded_mask_results
+        results.append(result)
+
         results.append(result)
 
         if rank == 0:
